TASK-6-8. The next changes are committed: 1. Cloud Tests (except Drag and Drop) 2. Logger 3. Making screenshots 4. Support chrome and firefox 5. Improvements for mail.ru tests
